refactor the profile into a benchmark
